Improve joypad feel by adjusting key clearing
The emulator key input still feels a bit clunky in Tetris, but that
might just be the Gameboy itself, not this implementation. For example,
if pressing arrow right, then arrow down, arrow down won't register even
if arrow right is unpressed as well. Mooneye-gb handles similarly,
so might not be fixable.