"Implement SampleBuf which is a conditional replacement for BlipBuf when using WASM"
The Rust Crate BlipBuf is currently used for sound downsampling and
is a wrapper over the C library BlipBuf. This gives great sound
quality, but breaks WASM support as it cannot be compiled into WASM.
SampleBuf has been implemented to be a drop-in replacement for BlipBuf
(with worse sound quality, as it currently only uses nearest neighbour
sampling). SampleBuf is then conditionally used for WASM compilation,
which restores core support for the wasm frontend.