Implement initial stages of a better sound syncing system
The previous method of skipping frames didn't work too well
and gave visible artifacts. By dynamically adjusting the
output rate slightly, we can dynamically make sure we
consume as much audio as we produce. This caused a change
in audio frequency, but with small enough values this is not
noticable to the human ear. The system is still a bit buggy
and untested.