Implement local caching of most recent savefile for web frontend
This allows users to easily continue from where they left off.
This is however quite slow, as Local Storage can only store
strings for some reason, so we have to convert the byte array
to a JSON object which takes some time and also means we take up
quite a bit of space on the user device. Maybe this can be done
better on the Rust side?