Fix issue where the first line after lcd enable was drawn blank.
This issue is related to the Gameboy hardware behaviour of drawing
a blank frame on the first frame after LCD enable. The emulator
does not currently do this, which caused an issue together with
the conditional drawing based on gpu state modification. The emulator now
instead draws two concurrent frames after a gpu modification. This
somewhat fixes the issue, although the first frame really should be blank
instead, which is a future project.